message_fetch: Move ok_to_include_history to zerver.lib.narrow.
